back/requirements.txt
Django==2.2.1
django-cors-headers==2.5.3
djangorestframework==3.9.3
pkg-resources==0.0.0
pytz==2019.1
sqlparse==0.3.0
